Implement integration tests and support windows
Previously, we run a kind of integration tests in the unit tests.
However, they are not enough and pretty slow because of Jenkins's
design of unit tests (always using new JENKINS_HOME by default).

This commit introduces completely new integration tests using real
Jenkins war file. With this method, we can reuse Jenkins plugins and
workspaces so that the tests are much faster.

At the same time, the unit tests are no more integration tests and
they use stub installer. Therefore, they are quick and we can run them
in parallel per class level.

Finally, we add Windows support if `bash` is available in the node.