Fix windows CI
The previous commit uses different node labels in `ensureInNode()` and
`buildPlugin()` so that there is no effect.

To make sure both stages run on the same node, we need to use only one
windows node and no container agent. (Actually, container agent can be
valid but the nodes look busy or unavailable for windows for now.)